test(core): migrate UpdateTabTests (Wave C)
65 tests, 165 failure sites preserved 1:1. The five confirmation-args
guards (`guard let confirmation = ... else { throw TestFailure }`) convert
to `Issue.record + return` -- `#require` cannot bind through the
case-pattern destructuring the dispatcher helpers do internally. File-
local `closeTabsConfirmationArgs` / `closeTabsConfirmationEffectCount` /
`effectCount` / `isTerminateEffect` helpers move alongside the suite.
